+++ frontend/src/router/routes/admin/settings.ts
@@ -4,22 +4,6 @@ import Permissions from '@/mythicaldash/Permissions';
 
 const settingsRoutes: RouteRecordRaw[] = [
     {
-        path: '/mc-admin/settings/old',
-        name: 'admin-settings',
-        component: () => import('@/views/admin/settings/Index.vue'),
-        meta: {
-            requiresAuth: true,
-            requiresAdmin: true,
-        },
-        beforeEnter: (to, from, next) => {
-            if (Session.hasOrRedirectToErrorPage(Permissions.ADMIN_SETTINGS_VIEW)) {
-                next();
-            } else {
-                next('/errors/403');
-            }
-        },
-    },
-    {
         path: '/mc-admin/settings',
         name: 'admin-settings-new',
         component: () => import('@/views/admin/settings/SettingsNew.vue'),